chore: rename utils.ts to fileoperations.ts
